Refactoring and documenting BTree
