+++ src/main/java/com/github/jambodb/storage/btrees/files/FileBTreePage.java
@@ -1,6 +1,8 @@
-package com.github.jambodb.storage.btrees;
+package com.github.jambodb.storage.btrees.files;
 
 import com.github.jambodb.storage.blocks.BlockStorage;
+import com.github.jambodb.storage.btrees.BTreePage;
+import com.github.jambodb.storage.btrees.Serializer;
 import com.github.jambodb.storage.utils.SerializableList;
 import java.io.IOException;
 import java.nio.ByteBuffer;
