+++ src/main/java/com/github/jambodb/storage/btrees/files/FilePager.java
@@ -1,7 +1,11 @@
-package com.github.jambodb.storage.btrees;
+package com.github.jambodb.storage.btrees.files;
 
 import com.github.jambodb.storage.blocks.BlockStorage;
 import com.github.jambodb.storage.blocks.FileBlockStorage;
+import com.github.jambodb.storage.btrees.BTreePage;
+import com.github.jambodb.storage.btrees.Pager;
+import com.github.jambodb.storage.btrees.Serializer;
+
 import java.io.IOException;
 import java.nio.BufferUnderflowException;
 import java.nio.file.Files;
