+++ src/test/java/com/github/jambodb/storage/btrees/FileBTreePageTests.java
@@ -2,6 +2,7 @@ package com.github.jambodb.storage.btrees;
 
 import com.github.jambodb.storage.blocks.BlockStorage;
 import com.github.jambodb.storage.blocks.FileBlockStorage;
+import com.github.jambodb.storage.btrees.files.FileBTreePage;
 import com.github.jambodb.storage.btrees.mock.MockObject;
 import com.github.jambodb.storage.btrees.mock.MockObjectSerializer;
 import com.github.jambodb.storage.btrees.mock.StringSerializer;
