+++ src/test/java/com/github/jambodb/storage/btrees/FilePagerTests.java
@@ -1,5 +1,7 @@
 package com.github.jambodb.storage.btrees;
 
+import com.github.jambodb.storage.btrees.files.FileBTreePage;
+import com.github.jambodb.storage.btrees.files.FilePager;
 import com.github.jambodb.storage.btrees.mock.MockObject;
 import com.github.jambodb.storage.btrees.mock.MockObjectSerializer;
 import com.github.jambodb.storage.btrees.mock.StringSerializer;
